requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
black==21.9b0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
filelock==3.3.1
flake8==4.0.1
fugashi==1.1.1
gitdb==4.0.8
GitPython==3.1.20
huggingface-hub==0.0.19
idna==3.3
importlib-metadata==4.2.0
importlib-resources==5.3.0
ipadic==1.0.0
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.0.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.0
ptyprocess==0.7.0
pyarrow==5.0.0
pycodestyle==2.8.0
pycparser==2.20
pydeck==0.6.2
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.10.23
requests==2.26.0
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.1.0
streamlit-player==0.1.5
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
torch==1.10.0
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
transformers==4.11.3
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.6.0
